refactor: managing TreeFolder conversion in UseCase
