+++ feature/navigation/drawer/src/debug/kotlin/app/k9mail/feature/navigation/drawer/ui/folder/FolderListPreview.kt
@@ -3,6 +3,7 @@ package app.k9mail.feature.navigation.drawer.ui.folder
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.PreviewWithTheme
+import app.k9mail.feature.navigation.drawer.domain.entity.TreeFolder
 import app.k9mail.feature.navigation.drawer.ui.FakeData.DISPLAY_FOLDER
 import app.k9mail.feature.navigation.drawer.ui.FakeData.UNIFIED_FOLDER
 import kotlinx.collections.immutable.persistentListOf
@@ -12,8 +13,8 @@ import kotlinx.collections.immutable.persistentListOf
 internal fun FolderListPreview() {
     PreviewWithTheme {
         FolderList(
-            folders = persistentListOf(
-                DISPLAY_FOLDER,
+            rootFolder = TreeFolder.createFromFolders(
+                persistentListOf(DISPLAY_FOLDER)
             ),
             selectedFolder = null,
             onFolderClick = {},
@@ -27,8 +28,8 @@ internal fun FolderListPreview() {
 internal fun FolderListPreviewSelected() {
     PreviewWithTheme {
         FolderList(
-            folders = persistentListOf(
-                DISPLAY_FOLDER,
+            rootFolder = TreeFolder.createFromFolders(
+                persistentListOf(DISPLAY_FOLDER)
             ),
             selectedFolder = DISPLAY_FOLDER,
             onFolderClick = {},
@@ -42,10 +43,10 @@ internal fun FolderListPreviewSelected() {
 internal fun FolderListWithUnifiedFolderPreview() {
     PreviewWithTheme {
         FolderList(
-            folders = persistentListOf(
+            rootFolder = TreeFolder.createFromFolders(persistentListOf(
                 UNIFIED_FOLDER,
                 DISPLAY_FOLDER,
-            ),
+            )),
             selectedFolder = DISPLAY_FOLDER,
             onFolderClick = {},
             showStarredCount = false,
