+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/NavigationDrawerModule.kt
@@ -6,6 +6,7 @@ import app.k9mail.feature.navigation.drawer.domain.DomainContract.UseCase
 import app.k9mail.feature.navigation.drawer.domain.usecase.GetDisplayAccounts
 import app.k9mail.feature.navigation.drawer.domain.usecase.GetDisplayFoldersForAccount
 import app.k9mail.feature.navigation.drawer.domain.usecase.GetDrawerConfig
+import app.k9mail.feature.navigation.drawer.domain.usecase.GetTreeFolders
 import app.k9mail.feature.navigation.drawer.domain.usecase.SaveDrawerConfig
 import app.k9mail.feature.navigation.drawer.domain.usecase.SyncAccount
 import app.k9mail.feature.navigation.drawer.domain.usecase.SyncAllAccounts
@@ -48,6 +49,10 @@ val navigationDrawerModule: Module = module {
         )
     }
 
+    single<UseCase.GetTreeFolders> {
+        GetTreeFolders()
+    }
+
     single<UseCase.SyncAccount> {
         SyncAccount(
             accountManager = get(),
@@ -67,6 +72,7 @@ val navigationDrawerModule: Module = module {
             saveDrawerConfig = get(),
             getDisplayAccounts = get(),
             getDisplayFoldersForAccount = get(),
+            getTreeFolders = get(),
             syncAccount = get(),
             syncAllAccounts = get(),
         )
