+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/domain/DomainContract.kt
@@ -5,6 +5,7 @@ import app.k9mail.feature.navigation.drawer.domain.entity.DisplayAccount
 import app.k9mail.feature.navigation.drawer.domain.entity.DisplayFolder
 import app.k9mail.feature.navigation.drawer.domain.entity.DisplayUnifiedFolder
 import app.k9mail.feature.navigation.drawer.domain.entity.DisplayUnifiedFolderType
+import app.k9mail.feature.navigation.drawer.domain.entity.TreeFolder
 import kotlinx.coroutines.flow.Flow
 
 internal interface DomainContract {
@@ -24,6 +25,10 @@ internal interface DomainContract {
 
         fun interface GetDisplayFoldersForAccount {
             operator fun invoke(accountId: String, includeUnifiedFolders: Boolean): Flow<List<DisplayFolder>>
+        }
+
+        fun interface GetTreeFolders {
+            operator fun invoke(folders: List<DisplayFolder>, maxDepth: Int): TreeFolder
         }
 
         /**
