+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/ui/DrawerContent.kt
@@ -64,7 +64,7 @@ internal fun DrawerContent(
                         .fillMaxSize(),
                 ) {
                     FolderList(
-                        folders = state.folders,
+                        rootFolder = state.rootFolder,
                         selectedFolder = state.folders.firstOrNull { it.id == state.selectedFolderId },
                         onFolderClick = { folder ->
                             onEvent(Event.OnFolderClick(folder))
